+++ src/components/orders-payments/OrderHistoryList/index.ts
@@ -7,3 +7,4 @@ export { default as OrderHistoryClientWrapper } from './OrderHistoryClientWrappe
 export { default as OrderDetailClientWrapper } from './OrderDetailClientWrapper';
 export { default as OrderDetailListGroupHeader } from './OrderDetailListGroupHeader';
 export { default as OrderGroupHeader } from './OrderGroupHeader';
+export { default as AllCancelButton } from './AllCancelButton';
